fix test_registration
