+++ tests/ui/guest/test_registration.py
@@ -540,5 +540,20 @@ class RegistrationTestCase(BaseTestRunner):
         self.assertEqual(result_notification.get_notification_text(), notification, f"Notification doesn't equal {notification}")
         sleep(10)
 
+    @allure.testcase('https://github.com/UA-1023-TAQC/SpaceToStudyTA/issues/189')
+    def test_visability_of_the_all_elements_after_resizing_for_who_we_are_block(self):
+        window_width = 600
+        window_height = 1000
+        who_we_are = (HomePageGuest(self.driver)
+                      .get_header()
+                      .get_navigate_links()[2]
+                      .click())
+        HomePageGuest(self.driver).set_size_window(window_width, window_height)
+        who_we_are_elements = (HomePageGuest(who_we_are)
+                               .get_who_we_are_block()
+                               .get_who_we_are_elements())
+        for element in who_we_are_elements.values():
+            self.assertTrue(element.is_displayed())
+
 if __name__ == '__main__':
     unittest.main(verbosity=2)
